# modulo_administracion_vehiculos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vehículos",
        "description": "Documentación de la API del sistema de administración de vehículos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/seguros": {
            "get": {
                "tags": [
                    "Seguro"
                ],
                "summary": "Lista todos los seguros",
                "description": "Obtiene todos los registros de seguros con su vehículo asociado",
                "operationId": "560a996edeccb88dec388fa203eb38f7",
                "responses": {
                    "200": {
                        "description": "Lista de seguros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Seguro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Seguro"
                ],
                "summary": "Crea un nuevo seguro",
                "description": "Crea un registro de seguro asociado a un vehículo existente",
                "operationId": "b4a9937038388bb0f9b60cc36b89e850",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehiculo_id",
                                    "tipo",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "vehiculo_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Seguro contra todo riesgo"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Seguro creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seguro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Vehículo no encontrado o datos inválidos"
                    }
                }
            }
        },
        "/api/seguros/{id}": {
            "get": {
                "tags": [
                    "Seguro"
                ],
                "summary": "Obtiene un seguro por ID",
                "description": "Devuelve un registro de seguro específico con su vehículo asociado",
                "operationId": "0862b6f58c6d0cd47bf9c484e5e04e2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del seguro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del seguro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seguro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Seguro no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Seguro"
                ],
                "summary": "Actualiza un seguro existente",
                "description": "Actualiza un registro de seguro por ID",
                "operationId": "9b0fd95798e599c7dd7b91d15b30d6a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del seguro a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehiculo_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Seguro a terceros"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seguro actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Seguro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Seguro no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Seguro"
                ],
                "summary": "Elimina un seguro",
                "description": "Elimina un registro de seguro por ID",
                "operationId": "235a20f5d5ee3ae9130b088da481736a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del seguro a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Seguro eliminado correctamente"
                    },
                    "404": {
                        "description": "Seguro no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Seguro": {
                "required": [
                    "vehiculo_id",
                    "compañía",
                    "tipo_cobertura",
                    "fecha_inicio",
                    "fecha_fin"
                ],
                "properties": {
                    "id_seguro": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehiculo_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "compañía": {
                        "type": "string",
                        "example": "Seguros Quito S.A."
                    },
                    "tipo_cobertura": {
                        "type": "string",
                        "example": "Todo riesgo"
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "fecha_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-01"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}